Adapt cmake scripts to work on linux x86_64 also.
Use cmake multi-architecture support to determine if we're not
building for windows, and if so, to use suitable options for linux
instead.  Ideally, this should support all the systems that are in
"build_detect_platform", but some of those look fairly exotic...

Some specific changes (for linux):
 use posix date command.
 run git directly.
 support out of tree build.
 different compiler options.
 different library paths.
 different portability sources.
Changes that should not break windows:
 completely different logic to make build_version.cc
 "make check" support.
 library ordering.

In linux (and many other unix systems), the link order of modules
is critical.  It's necessary to list libraries so that at least one
reference occurs before the module that satisfies that reference.
So testutillib must be supplied first, before rocksdb on the link line.
For the case of gtest, it's necessary this be supplied last, after all
references, otherwise bad things happen (like every test memory faults.)
I believe this stricter ordering should also work on aix, windows,
and most other systems.

Further oddities:
When building the shared library for rocksdb, the logic was including
LIBS which pulled in both the static library for rocksdb, and gtest.
That explains why I needed -fPIC for gtest, but the better fix is
not to include an extra spurious copy of gtest in the rocksdb library.
This fix is absolutely necessary for rpmbuild, because that passes in
its own flags to cmake to tell it to build everything as shared libraries.

I still need to link the shared rocksdb with the third party dependencies,
excepting not gtest.  Also I need to add gtest explicitly for ctest.

Signed-off-by: Marcus Watts <mwatts@redhat.com>